==> 00_env_init.Rmd <==
---
title: "00_env_init"
author: "Jagadeesh Puvvula"
date: "2023-08-02"
output: pdf_document
---

#library and r env
```{r}
library(pacman)
pacman::p_load(tidyverse, janitor, reshape2, haven, qgcomp, qgcompint, bkmr)

#load these libraries from below to get the versions while wrapping up the project
#pacman::p_load(qgcomp, bkmr, Matrix, fields, bindrcpp, gglasso, grpreg, sparsegl, Matrix, glmnet, stabs)

# Call the function to print the version information
#print_version_info()
```

#folder locations
```{r}
data <- "E:/BBK17/pj/basc_brief_analy/data/"
result<- "E:/BBK17/pj/basc_brief_analy/result/"
analy_dat<- "E:/BBK17/pj/basc_brief_analy/analy_dat/"
```
