==> 01_data_prep.Rmd <==
---
title: "01_data_prep"
author: "Jagadeesh Puvvula"
date: "2023-07-21"
output: pdf_document
---

#Select variables for BASC and BRIEF analysis 
```{r}
dat_origin<-  read_csv("E:/BBK17/pj/data_2023apr/obs_w_outcome/home_mirec_w_out.csv") |>
  mutate(pb = if_else(cohort == "1", pb * 10, pb))

dat_basc<- dat_origin |>
  clean_names()|>
  select(-c(43:45))|>
  rename(Pb=pb, Hg=hg, DMA=dma, DDE= dde, PBDE_47 = pbde47,
         PCB_118=pcb118, PCB_138=pcb138, PCB_153=pcb153,
         PCB_180=pcb180, PFHxS=pfhxs, PFOA=pfoa, PFOS=pfos,
         BCEtP=bcetp, BDCPP=bdcpp, DBuP=dbup, DPhP=dphp,
         TCS=tcs, BPA=bpa, MBP=mbp, MBZP=mbzp, MCPP=mcpp,
         sigma_DEHP=dehp, MEP=mep, MIBP=mibp, di_Ethyl_OP=op_de,
         di_Methyl_OP=op_dm, B_PB=b_pb, M_PB=m_pb, P_PB=p_pb, city=center,
         mom_edu_cat=edu3)|>
  mutate(race_bin = if_else(race == "white", 1, 2),
         sex = if_else(sex == "Male", 1, 2)) |>
  filter(if_all(43:47, ~ !is.na(.) & !grepl("^\\s*$", .) & !grepl("^\\.$", .))) |> #if_any didnt work here
  filter(basc_ext != 0 & basc_inz != 0 & basc_bsi != 0 & brief_wm != 0 & brief_po != 0) |>
  left_join(mat_dep, by="subject_id")
```

#missing data viz - BASC/BRIEF
```{r}
library(naniar)

vis_miss(dat_basc[,4:48], cluster = T)
```

#count observations if any of chemical biomarker is present
```{r}
any_exp_cnt<- dat_basc |>
  filter(!if_all(4:32, is.na))

missing_home_score<- dat_basc |>
  filter(subject_id %in% dat$subject_id)
```


# export data version for behavioral outcome analysis (all obs with outcome - has missing exposures)
```{r}
write_csv(dat_basc, paste0(data, "dat_basc.csv"))
```

#get subset with all outcomes and exposures
```{r}
dat_basc<- read_csv(paste0(data, "dat_basc.csv"))

dat_basc_all_exp<- dat_basc |>
  filter(if_all(c(4:32), ~ !is.na(.) & !grepl("^\\s*$", .) & !grepl("^\\.$", .))) |>
  mutate(mom_edu_cat = if_else(is.na(mom_edu_cat), 2, mom_edu_cat)) |> # replaced with category-2 (dominant)
  select(-c("b_length", "b_wght", "race"))


# 100 obs missing HOME score
x<- dat_basc_all_exp |>
  group_by(cohort, sex, city) |>
  summarise(missing_home = sum(is.na(home_score_total)))
```

#missing data summary
```{r}
dat_miss<- missing_data_summary(dat_basc_all_exp) #ref999
```

#Impute unmeasured HOME score data using SuperLearner
```{r}
pred_var<- "home_score_total"
dat_pred_v1 <- impute_missing(dat_basc_all_exp, {{pred_var}})

#filter unknown data for predictions
pred_set <- dat_pred_v1 |>
  filter(is.na(dat_pred_v1[[pred_var]])) |>
  select(-c(2:3))


#for training the prediction model
dat_n <- dat_pred_v1 |> drop_na() |>
  select(-c(1:3))

set.seed(123, "L'Ecuyer-CMRG")
indx<- sample(nrow(dat_n), round(0.75*nrow(dat_n)))
train<- dat_n[indx,]
test<- dat_n[-indx,]

y_train<- train[[pred_var]]
x_train<- train[, -which(colnames(train) == pred_var)]

x_test <- as.matrix(test[, -which(colnames(test) == pred_var)])
y_test <- as.matrix(test[[pred_var]])

#model object
sl.obj2<- SuperLearner(y_train, x_train, family = gaussian(), 
                       SL.library = c("SL.mean", "SL.xgboost", "SL.gbm"))
#model validation to get MSE
validation<- predict(sl.obj2, test[, -which(colnames(test) == pred_var)])


#get predictions: Input - > variable column number being predicted, variable name and rounding
x <- generate_x(sl.obj = sl.obj2, pred_set = pred_set, id_col = 1, 
                pred_var = pred_var, rounding = 1)


#replaces NA values with the predicted values
original <- replace_values(x, dat_basc_all_exp)
```

#export dataset with inputed HOME score - no more NA values 
```{r}
write_csv(original, paste0(data, "basc_imputed.csv"))
```

#adding maternal depression variable
```{r}
df<- read_csv(paste0(data, "basc_imputed.csv"))

#BDI Score baseline - calculated z-scores
home_mat_dep<- read_sas(paste0(data, "HOME_covariates.sas7bdat")) |>
  clean_names() |>
  select(c(1,23))|>
  mutate(subject_id = as.character(subject_id))|>
  rename(mat_dep = bdi_bhv) |>
  mutate(mat_dep_z = (mat_dep - mean(mat_dep, na.rm = TRUE)) / sd(mat_dep, na.rm = TRUE))

#CES-D score from MIREC - calcualted z-scores
mirec_mat_dep<- read_csv("E:\\BBK17\\4. BBK17_CD+ CRF data corrections_Nov18 2022\\CDPLUS.2_NEURO.7_20221117\\MSAQ_DATA_DEPRESSION_Yes to NEURO BBK.csv") |>
  clean_names() |>
  select(c(2,20)) |>
  rename(mat_dep = cesdepq5cd)|>
  mutate(mat_dep_z = (mat_dep - mean(mat_dep, na.rm = TRUE)) / sd(mat_dep, na.rm = TRUE))


#join maternal depression data from HOME and MIREC
mat_dep<- bind_rows(home_mat_dep, mirec_mat_dep) |>
  select(c(1,3))

df_nov_2024<- left_join(df, mat_dep, by="subject_id") |>
  mutate(mat_dep_z = if_else(is.na(mat_dep_z), 
                             median(mat_dep_z, na.rm = TRUE), 
                             mat_dep_z))

write_csv(df_nov_2024, paste0(data, "basc_imputed_nov_2024.csv"))
```

#correlations
#load dat from 04_ee_qgcomp file
```{r}
dat_cor<- dat |> select(c(1:29, 31,32, 39,45,40:44)) |> #38 = HOME score
  drop_na() |>
  rename_with(~ gsub("_", "-", toupper(.)))

cormat <- round(x=cor(dat_cor, method = "spearman", use = "complete.obs"), digits = 2) |>
  melt() |> 
  clean_names()

ggplot(cormat, aes(x = var2, y = var1, fill = value)) + #, label = value
  geom_tile(color = "white") +
  #geom_text(color = "black", size = 3, vjust = 1) +
  scale_fill_gradient2(low = "red", high = "blue", mid = "white",
                       midpoint = 0,
                       limit = c(-1, 1), space = "Lab",
                       name = "Spearman Correlation | HOME & MIREC [n=695]") +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, size = 11, hjust = 1),
        axis.text.y = element_text(angle = 0, vjust = 0.5, size = 11, hjust = 1),
        axis.title.x = element_blank(),
        axis.title.y = element_blank(),
        panel.grid.major = element_blank(),
        panel.border = element_blank(),
        panel.background = element_blank(),
        axis.ticks = element_blank(),
        legend.position = "bottom", legend.box = "horizontal") +
  coord_fixed() +
  # add vertical and horizontal lines to separate variable groups
  geom_vline(xintercept = c(0.5, 29.5, 33.5, 38.5), color = "black", size=1.25) +
  geom_hline(yintercept = c(0.5, 29.5, 33.5, 38.5), color = "black", size=1.25) 
```

#export figure
```{r}
ggsave("E:/BBK17/pj/basc_brief_analy/result/spearman_corr.tiff", 
       width = 12,height = 12,
       dpi=300)
```
