==> 04.1_qgc_wo_mat_dep.Rmd <==
---
title: "04.1_qgc_without_mat_dep"
author: "Jagadeesh Puvvula"
date: "2025-02-05"
output: pdf_document
---

#for all 29 chemicals
```{r}
dat_si <- read_csv(paste0(data, "basc_imputed_nov_2024.csv")) 

dat<- dat_si |>
  rename(BCEP=BCEtP, BDCIPP=BDCPP, DNBP=DBuP, DPhP=DPhP)|>
  mutate(across(all_of(c("cohort", "city", "sex", "race_bin", 
                         "parity_n", "mom_edu_cat")), as.factor),
         sex = ifelse(sex == "1", "Male", "Female"),
         log_cotinine = log2(cotinine+ 1e-6)) |>
  filter(basc_ext != 0 & basc_inz != 0 & basc_bsi != 0 & brief_wm != 0 & brief_po != 0) |>
  select(c(Pb, Hg, DMA, DDE, PBDE_47, PCB_118, PCB_138, PCB_153, PCB_180,
          PFHxS, PFOA, PFOS, BCEP, BDCIPP, DNBP, DPhP, TCS, BPA, MBP, MBZP,
          MCPP, sigma_DEHP, MEP, MIBP, di_Ethyl_OP, di_Methyl_OP, 
          B_PB, M_PB, P_PB, log_cotinine, home_score_total, mom_age, 
          cohort, city, sex, race_bin,  mom_edu_cat, parity_n, 
          basc_ext, basc_inz, basc_bsi, brief_wm, brief_po)) 

mixture<- names(dat[c(1:29)])
```

#Run models
```{r}
qgcomp_func(outcomes = c("basc_ext", "basc_inz", "basc_bsi", "brief_wm", "brief_po"), 
            data = dat,
            output_folder = "E:/BBK17/pj/basc_brief_analy/result/qgcomp/wo_mat_dep", 
            include_sex = TRUE, 
            include_cohort = TRUE, 
            chemicals = list("Pb", "Hg", "DMA", 
                             "DDE", "PBDE_47", 
                             "PCB_118", "PCB_138", "PCB_153", "PCB_180", 
                             "PFHxS", "PFOA", "PFOS", 
                             "BCEP", "BDCIPP", "DNBP", "DPhP", 
                             "TCS", "BPA", 
                             "MBP", "MBZP", "MCPP", "sigma_DEHP", 
                             "MEP", "MIBP", 
                             "di_Ethyl_OP", "di_Methyl_OP", 
                             "B_PB", "M_PB", "P_PB"),
            covariates = list("race_bin", "log_cotinine", "mom_edu_cat", 
                              "home_score_total", "parity_n", "mom_age"),
            q = 4, b=400, seed = 2025)
```

#extract results
```{r}
results<- get_gcomp_estimates(results_folder = "E:/BBK17/pj/basc_brief_analy/result/qgcomp/wo_mat_dep")

write_csv(results, "E:/BBK17/pj/basc_brief_analy/result/qgcomp/wo_mat_dep/r_obj_summ.csv")
```

#data prep for visualizing estimates and CI
```{r}
res<- results |>
  rename(bootstrap=boot_strp) |>
  mutate(p_value = if_else(p_value <= 0.05, "<0.05", ">0.05"),
         ci_null = (lower_ci <=0 & upper_ci >=0)) |>
  mutate(cohort = fct_recode(as.factor(cohort), Pooled = "all", HOME = "home", MIREC = "mirec"),
         gender = fct_recode(as.factor(gender), All = "all", Female = "Female", Male = "Male"),
         bootstrap = fct_recode(as.factor(bootstrap), Yes = "boot", No = "nb"),
         outcome = fct_recode(as.factor(outcome), BSI = "bsi", EXT = "ext", INZ = "inz",
                              PO = "po", WM = "wm"),
         outcome_cat = case_when(outcome %in% c("BSI", "EXT", "INZ") ~ "BASC",
                                 TRUE ~ "BRIEF")) |>
  filter(abs(lower_ci) <= 40 & abs(upper_ci) <= 40,
         cohort == "Pooled")

res$outcome_cat<- factor(res$outcome_cat, levels = c("BASC", "BRIEF"))
res$bootstrap<- factor(res$bootstrap, levels = c("Yes", "No"))

cbbPalette <- c("black", "#0072B2")
```

```{r}
ggplot(res, aes(x = as.factor(outcome), y = estimate, ymin = lower_ci, ymax = upper_ci)) + 
  geom_pointrange(aes(col = as.factor(ci_null), shape=bootstrap), 
                  position=position_dodge(width=0.5),size = 0.6) +
  geom_hline(aes(yintercept = 0), linetype="dashed") + 
  scale_colour_manual(values=cbbPalette) +
  labs(x=NULL,
       y=expression(paste(beta[Q-gcomp]," [95% CI]")),
       title = "",
       caption = "")+
  theme_bw()+
  theme(axis.line = element_line(colour = "black"))+
  theme(plot.caption = element_text(size = 6),
        plot.title=element_text(size = 10, hjust = 0),
        panel.border = element_blank(),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        axis.line = element_line(colour = "black"),
        text=element_text(size=10),
        axis.text = element_text(size = 10),
        axis.text.x=element_text(size=10,angle=0),
        axis.ticks.x=element_blank(),
        strip.text.x=element_text(size=10,angle=0),
        strip.text.y = element_text(size=10),
        strip.placement = "outside",
        strip.background = element_rect(fill=NULL, colour = NULL),
        legend.box.just = "center",
        legend.spacing = unit(0, "cm"),
        legend.position = "bottom",
        legend.box = "horizontal",
        legend.direction = "horizontal",
        legend.text = element_text(size = 8),
        legend.key = element_blank(),
        legend.key.height = unit(3, "mm"))+
  guides(col="none",
         shape=guide_legend(title = "Bootstrap"))+
  facet_grid(gender~outcome_cat, scales = "free", switch = "x", space = "free_x")
```

#export figure
```{r}
ggsave("E:/BBK17/pj/basc_brief_analy/result/qgcomp/wo_mat_dep/qgcomp_29chem.tiff", 
       width = 8,height = 6,
       dpi=300)
```

#extract chemical weights 
```{r}
qgcomp_weights <- extract_weights("E:/BBK17/pj/basc_brief_analy/result/qgcomp/wo_mat_dep")
write_csv(qgcomp_weights, "E:/BBK17/pj/basc_brief_analy/result/qgcomp/wo_mat_dep/chem_weight.csv")
```

# load chemical weights
```{r}
qgcomp_weights<- read_csv("E:/BBK17/pj/basc_brief_analy/result/qgcomp/wo_mat_dep/chem_weight.csv")
```


#main result
```{r}
results<- read_csv("E:/BBK17/pj/basc_brief_analy/result/qgcomp/wo_mat_dep/r_obj_summ.csv") |>
  filter(cohort == "all" ) |>
  mutate(result = sprintf("%.2f (%.2f, %.2f)", estimate, lower_ci, upper_ci)) |>
  mutate(gender_outcome = paste(gender, outcome, sep = "_")) |>
  select(c(1,10,9)) |>
  pivot_wider(names_from = boot_strp, values_from = result)
```

#print a table using qgcomp results
```{r}
custom_order <- c("PSI1", "NEG.PSI", "POS.PSI", "PB", "HG", "DMA", "DDE", "PBDE47", "PCB118", "PCB138", "PCB153", "PCB180",
                  "PFHXS", "PFOA", "PFOS", "BCEP", "BDCIPP", "DNBP", "DPHP",
                  "TCS", "BPA", "MBP", "MBZP", "MCPP", "SIGMA DEHP", "MEP", "MIBP",
                  "DI ETHYL OP", "DI METHYL OP", "B PB", "M PB", "P PB")

dat <- qgcomp_weights |> 
  mutate(file_name = str_remove(file_name, "nb_")) |>
  separate(file_name, c("gender", "cohort", "outc_cat", "outcome"), sep = "_") |>
  mutate(weight= if_else(direction=="neg", weight*-1, weight),
         chemical=gsub("log_", "", chemical),
         outcome = str_to_upper(outcome),
         cohort = fct_recode(cohort, "Pooled" = "all",
                             "HOME" = "home",
                             "MIREC" = "mirec"),
         gender =fct_recode(gender, "All"= "all"),
         chemical = toupper(str_replace_all(chemical, "_(?=[A-Z])", " ")),
         chemical = str_replace_all(chemical, "_(?=[0-9])", ""),
         chem_grp = ifelse(grepl("psi", direction), "psi", "chem")) |>
  filter(cohort== "Pooled")

dat_tbl <- dat |>
  mutate(chemical = factor(chemical, levels = custom_order)) |>
  arrange(chemical) |>
  select(-c(3,6,8)) |>
  pivot_wider(names_from = c("gender", "cohort", "outcome"), values_from = "weight") |>
  select(
    chemical,
    starts_with("All") & contains("BSI"),
    starts_with("Female") & contains("BSI"),
    starts_with("Male") & contains("BSI"),
    
    starts_with("All") & contains("EXT"),
    starts_with("Female") & contains("EXT"),
    starts_with("Male") & contains("EXT"),
    
    starts_with("All") & contains("INZ"),
    starts_with("Female") & contains("INZ"),
    starts_with("Male") & contains("INZ"),
    
    starts_with("All") & contains("PO"),
    starts_with("Female") & contains("PO"),
    starts_with("Male") & contains("PO"),
    
    starts_with("All") & contains("WM"),
    starts_with("Female") & contains("WM"),
    starts_with("Male") & contains("WM"),
  ) |>
  mutate_if(is.numeric, ~ as.numeric(sprintf("%.2f", .))) |>
  rename_all(~sub("_Pooled", "", .))|>
  mutate_at(vars(-1), ~ ifelse(is.na(.), "--", .))

write_csv(dat_tbl, "E:/BBK17/pj/basc_brief_analy/result/qgcomp/wo_mat_dep/gcomp_subset_weight.csv")

```
